requirements.txt
numpy==1.18.1
Pillow==7.1.2
scikit-learn==0.21.3
torch==1.6.0
torchvision==0.7.0